+++ server/infra/resource/src/database/forms.rs
@@ -4,4 +4,3 @@ pub mod comment;
 pub mod form;
 pub mod form_label;
 pub mod message;
-pub mod question;
